tmp3
tmp2

tmp

DecodeTypeRegister - ConfigureTypeRegister

one phase DecodeTypeImmediate().

Print number of executed instructions.

Temp speed improvement for InstructionType()

32: accessors + friend

Member param struct 32 bit.

Member param struct.

Sim debug.

BUG=

patch from issue 1310443004 at patchset 60001 (http://crrev.com/1310443004#ps60001)

+++ src/arm/simulator-arm.h
@@ -390,7 +390,7 @@ class Simulator {
   // Simulator support.
   char* stack_;
   bool pc_modified_;
-  int icount_;
+  uint64_t icount_;
 
   // Debugger input.
   char* last_debugger_input_;
